# public/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "JobBoard API",
    "description": "Production-grade Job Board REST API built with Laravel, Sanctum, Redis and RBAC.",
    "contact": { "email": "ravi194455@gmail.com", "name": "Ravi Kumar" },
    "version": "1.0.0"
  },
  "servers": [{ "url": "http://localhost:8000/api", "description": "Local Development" }],
  "paths": {
    "/auth/register": {
      "post": {
        "tags": ["Auth"], "summary": "Register a new user",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "required": ["name","email","password","password_confirmation","role_id"], "properties": { "name": {"type":"string","example":"Ravi Kumar"}, "email": {"type":"string","example":"ravi@example.com"}, "password": {"type":"string","example":"password"}, "password_confirmation": {"type":"string","example":"password"}, "role_id": {"type":"integer","example":3,"description":"2=Employer, 3=Candidate"} } } } } },
        "responses": { "201": {"description":"Registration successful"}, "422": {"description":"Validation error"} }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Auth"], "summary": "Login and get token",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "required": ["email","password"], "properties": { "email": {"type":"string","example":"candidate@jobboard.com"}, "password": {"type":"string","example":"password"} } } } } },
        "responses": { "200": {"description":"Login successful"}, "422": {"description":"Invalid credentials"} }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": ["Auth"], "summary": "Logout user",
        "security": [{"sanctum":[]}],
        "responses": { "200": {"description":"Logged out"}, "401": {"description":"Unauthenticated"} }
      }
    },
    "/profile": {
      "get": {
        "tags": ["Auth"], "summary": "Get authenticated user profile",
        "security": [{"sanctum":[]}],
        "responses": { "200": {"description":"Profile data"}, "401": {"description":"Unauthenticated"} }
      }
    },
    "/jobs": {
      "get": {
        "tags": ["Jobs"], "summary": "List all open jobs with filters",
        "parameters": [
          {"name":"location","in":"query","schema":{"type":"string"},"description":"Filter by location"},
          {"name":"type","in":"query","schema":{"type":"string","enum":["full-time","part-time","remote","contract"]}},
          {"name":"category_id","in":"query","schema":{"type":"integer"}},
          {"name":"search","in":"query","schema":{"type":"string"}}
        ],
        "responses": { "200": {"description":"Paginated job listings"} }
      },
      "post": {
        "tags": ["Jobs"], "summary": "Create a job (Employer only)",
        "security": [{"sanctum":[]}],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "required": ["category_id","title","description","location","type"], "properties": { "category_id": {"type":"integer","example":1}, "title": {"type":"string","example":"Senior PHP Developer"}, "description": {"type":"string","example":"We need a strong PHP developer"}, "location": {"type":"string","example":"Bangalore"}, "type": {"type":"string","enum":["full-time","part-time","remote","contract"],"example":"full-time"}, "salary_min": {"type":"number","example":800000}, "salary_max": {"type":"number","example":1200000} } } } } },
        "responses": { "201": {"description":"Job created"}, "403": {"description":"Unauthorized - Employer only"} }
      }
    },
    "/jobs/{id}": {
      "get": {
        "tags": ["Jobs"], "summary": "Get a single job",
        "parameters": [{"name":"id","in":"path","required":true,"schema":{"type":"integer"}}],
        "responses": { "200": {"description":"Job details"}, "404": {"description":"Not found"} }
      },
      "put": {
        "tags": ["Jobs"], "summary": "Update a job (Employer only)",
        "security": [{"sanctum":[]}],
        "parameters": [{"name":"id","in":"path","required":true,"schema":{"type":"integer"}}],
        "requestBody": { "content": { "application/json": { "schema": { "properties": { "title": {"type":"string"}, "description": {"type":"string"}, "location": {"type":"string"}, "type": {"type":"string","enum":["full-time","part-time","remote","contract"]}, "status": {"type":"string","enum":["open","closed"]}, "salary_min": {"type":"number"}, "salary_max": {"type":"number"} } } } } },
        "responses": { "200": {"description":"Job updated"}, "403": {"description":"Unauthorized"} }
      },
      "delete": {
        "tags": ["Jobs"], "summary": "Delete a job (Employer only)",
        "security": [{"sanctum":[]}],
        "parameters": [{"name":"id","in":"path","required":true,"schema":{"type":"integer"}}],
        "responses": { "200": {"description":"Job deleted"}, "403": {"description":"Unauthorized"} }
      }
    },
    "/my-jobs": {
      "get": {
        "tags": ["Jobs"], "summary": "Get employer's own jobs",
        "security": [{"sanctum":[]}],
        "responses": { "200": {"description":"Paginated list of employer jobs"} }
      }
    },
    "/jobs/{jobId}/apply": {
      "post": {
        "tags": ["Applications"], "summary": "Apply for a job (Candidate only)",
        "security": [{"sanctum":[]}],
        "parameters": [{"name":"jobId","in":"path","required":true,"schema":{"type":"integer"}}],
        "requestBody": { "content": { "application/json": { "schema": { "properties": { "cover_letter": {"type":"string","example":"I am a strong PHP developer with 2 years experience."}, "resume_url": {"type":"string","example":"https://ravi-kr.netlify.app"} } } } } },
        "responses": { "201": {"description":"Application submitted"}, "422": {"description":"Already applied or job closed"}, "403": {"description":"Candidates only"} }
      }
    },
    "/my-applications": {
      "get": {
        "tags": ["Applications"], "summary": "Get my applications (Candidate only)",
        "security": [{"sanctum":[]}],
        "responses": { "200": {"description":"Paginated list of applications"} }
      }
    },
    "/applications/{id}": {
      "delete": {
        "tags": ["Applications"], "summary": "Withdraw a pending application (Candidate only)",
        "security": [{"sanctum":[]}],
        "parameters": [{"name":"id","in":"path","required":true,"schema":{"type":"integer"}}],
        "responses": { "200": {"description":"Application withdrawn"}, "422": {"description":"Cannot withdraw non-pending application"} }
      }
    },
    "/jobs/{jobId}/applications": {
      "get": {
        "tags": ["Applications"], "summary": "View applications for a job (Employer only)",
        "security": [{"sanctum":[]}],
        "parameters": [{"name":"jobId","in":"path","required":true,"schema":{"type":"integer"}}],
        "responses": { "200": {"description":"Paginated list of applications"} }
      }
    },
    "/applications/{id}/status": {
      "put": {
        "tags": ["Applications"], "summary": "Update application status (Employer only)",
        "security": [{"sanctum":[]}],
        "parameters": [{"name":"id","in":"path","required":true,"schema":{"type":"integer"}}],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "required": ["status"], "properties": { "status": {"type":"string","enum":["reviewed","accepted","rejected"],"example":"accepted"} } } } } },
        "responses": { "200": {"description":"Status updated"}, "403": {"description":"Unauthorized"} }
      }
    },
    "/categories": {
      "get": {
        "tags": ["Categories"], "summary": "List all categories with job count",
        "responses": { "200": {"description":"List of categories"} }
      },
      "post": {
        "tags": ["Categories"], "summary": "Create a category (Admin only)",
        "security": [{"sanctum":[]}],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "required": ["name"], "properties": { "name": {"type":"string","example":"Backend Development"} } } } } },
        "responses": { "201": {"description":"Category created"}, "403": {"description":"Admin only"} }
      }
    },
    "/categories/{id}": {
      "delete": {
        "tags": ["Categories"], "summary": "Delete a category (Admin only)",
        "security": [{"sanctum":[]}],
        "parameters": [{"name":"id","in":"path","required":true,"schema":{"type":"integer"}}],
        "responses": { "200": {"description":"Category deleted"}, "403": {"description":"Admin only"} }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "sanctum": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your token: Bearer {token}"
      }
    }
  },
  "tags": [
    {"name":"Auth","description":"Authentication endpoints"},
    {"name":"Jobs","description":"Job listing endpoints"},
    {"name":"Applications","description":"Job application endpoints"},
    {"name":"Categories","description":"Category endpoints"}
  ]
}